+++ src/OpenTelemetry.Api/Logs/Logger.cs
@@ -48,6 +48,13 @@ abstract class Logger
     /// </summary>
     public string? Version { get; private set; }
 
+#pragma warning disable CA1056 // URI-like properties should not be strings
+    /// <summary>
+    /// Gets the schema URL of the instrumentation library.
+    /// </summary>
+    public string? SchemaUrl { get; private set; }
+#pragma warning restore CA1056 // URI-like properties should not be strings
+
     /// <summary>
     /// Emit a log.
     /// </summary>
@@ -64,5 +71,9 @@ abstract class Logger
         in LogRecordData data,
         in LogRecordAttributeList attributes);
 
-    internal void SetInstrumentationScope(string? version) => this.Version = version;
+    internal void SetInstrumentationScope(string? version, string? schemaUrl)
+    {
+        this.Version = version;
+        this.SchemaUrl = schemaUrl;
+    }
 }