+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -15,6 +15,10 @@ Notes](../../RELEASENOTES.md).
 * Reduced allocations when constructing `SpanAttributes` from an array.
   ([#7698](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7698))
 
+* Added `Logger.SchemaUrl` and a `LoggerProvider.GetLogger` overload accepting
+  a schema URL parameter using the new `LoggerOptions` type.
+  ([#7727](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7727))
+
 ## 1.18.0
 
 Released 2026-Aug-21